+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.0.7
+* add smart status widget
+## 0.0.6
+
+* add separated widget
 ## 0.0.5
 
 * add context extension